+++ CHANGELOG.md
@@ -3,7 +3,18 @@
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
-and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+and this project adheres to [Semantic Versioning](https://semver.org/spec2.0.0.html).
+
+## v1.2.2 - Nov 15, 2025
+
+### Added
+
+- Add unitsData in `questsData.ts`
+
+### Changed
+
+- Update `questlines.ts` to handle units
+- Replace `heavy` by `heavyinfantry` in `questlines.ts`
 
 ## v1.2.1 - Nov 10, 2025